# api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Fiber Practice API with OpenTelemetry and Swagger documentation",
        "title": "GFP API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Returns a simple greeting message",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Hello, World!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/simple-return/nested/{user}": {
            "get": {
                "description": "Returns a user object with ID, name, and personalized greeting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Nested user return",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User object with greeting",
                        "schema": {
                            "$ref": "#/definitions/nested.User"
                        }
                    }
                }
            }
        },
        "/simple-return/{user}": {
            "get": {
                "description": "Returns a greeting text for the specified user",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Simple text return",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Greeting text",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{user}": {
            "get": {
                "description": "Returns a personalized greeting for the specified user",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user greeting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personalized greeting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "nested.User": {
            "type": "object",
            "properties": {
                "greetingText": {
                    "type": "string",
                    "example": "Hello, alice!"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "alice"
                }
            }
        }
    }
}